refactor: move personal-workbench into admin
