+++ modules/apps/admin/personal-workbench/component-protocol/components/personal-workbench/messageList/messageList.go
@@ -33,10 +33,10 @@ import (
 	"github.com/erda-project/erda-infra/providers/component-protocol/utils/cputil"
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/bundle"
-	"github.com/erda-project/erda/modules/apps/personal-workbench/component-protocol/components/personal-workbench/common"
-	"github.com/erda-project/erda/modules/apps/personal-workbench/component-protocol/components/personal-workbench/common/gshelper"
-	"github.com/erda-project/erda/modules/apps/personal-workbench/component-protocol/types"
-	"github.com/erda-project/erda/modules/apps/personal-workbench/services/workbench"
+	"github.com/erda-project/erda/modules/apps/admin/personal-workbench/component-protocol/components/personal-workbench/common"
+	"github.com/erda-project/erda/modules/apps/admin/personal-workbench/component-protocol/components/personal-workbench/common/gshelper"
+	"github.com/erda-project/erda/modules/apps/admin/personal-workbench/component-protocol/types"
+	"github.com/erda-project/erda/modules/apps/admin/personal-workbench/services/workbench"
 	"github.com/erda-project/erda/pkg/strutil"
 	rt "github.com/erda-project/erda/pkg/time/readable_time"
 )
