+++ modules/apps/admin/personal-workbench/component-protocol/components/personal-workbench/workCards/provider.go
@@ -27,10 +27,10 @@ import (
 	"github.com/erda-project/erda-infra/providers/component-protocol/cptype"
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/bundle"
-	"github.com/erda-project/erda/modules/apps/personal-workbench/component-protocol/components/personal-workbench/common"
-	"github.com/erda-project/erda/modules/apps/personal-workbench/component-protocol/components/personal-workbench/i18n"
-	"github.com/erda-project/erda/modules/apps/personal-workbench/component-protocol/types"
-	"github.com/erda-project/erda/modules/apps/personal-workbench/services/workbench"
+	"github.com/erda-project/erda/modules/apps/admin/personal-workbench/component-protocol/components/personal-workbench/common"
+	"github.com/erda-project/erda/modules/apps/admin/personal-workbench/component-protocol/components/personal-workbench/i18n"
+	"github.com/erda-project/erda/modules/apps/admin/personal-workbench/component-protocol/types"
+	"github.com/erda-project/erda/modules/apps/admin/personal-workbench/services/workbench"
 )
 
 type WorkCards struct {
