+++ modules/apps/admin/personal-workbench/component-protocol/components/personal-workbench/workList/project.go
@@ -23,9 +23,9 @@ import (
 	"github.com/erda-project/erda-infra/providers/component-protocol/cptype"
 	"github.com/erda-project/erda-infra/providers/component-protocol/utils/cputil"
 	"github.com/erda-project/erda/apistructs"
-	"github.com/erda-project/erda/modules/apps/personal-workbench/component-protocol/components/personal-workbench/common"
-	"github.com/erda-project/erda/modules/apps/personal-workbench/component-protocol/components/personal-workbench/i18n"
-	wb "github.com/erda-project/erda/modules/apps/personal-workbench/services/workbench"
+	"github.com/erda-project/erda/modules/apps/admin/personal-workbench/component-protocol/components/personal-workbench/common"
+	"github.com/erda-project/erda/modules/apps/admin/personal-workbench/component-protocol/components/personal-workbench/i18n"
+	wb "github.com/erda-project/erda/modules/apps/admin/personal-workbench/services/workbench"
 )
 
 // GenProjKvColumnInfo show type: DevOps, MSP, DevOps(primary)/MSP
