+++ modules/apps/admin/personal-workbench/component-protocol/scenarios/personal-workbench.yml
