+++ modules/apps/admin/personal-workbench/i18n/files/scenarios/personal-workbench.yaml
